% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_shinybrms.R
\name{launch_shinybrms}
\alias{launch_shinybrms}
\title{Launch the \pkg{shinybrms} Shiny app}
\usage{
launch_shinybrms(...)
}
\arguments{
\item{\dots}{Arguments passed to \code{\link[shiny:runApp]{shiny::runApp()}}. For example, in order to force launching the
Shiny app in the system's default web browser (even if using RStudio), set argument
\code{launch.browser} to \code{TRUE}.}
}
\value{
\code{TRUE} (invisible).
}
\description{
This function launches the \pkg{shinybrms} Shiny app.
}
\note{
The following global options are introduced by \pkg{shinybrms}:
\itemize{
\item \code{shinybrms.prog_browser}: The value for the global option \code{browser} to be used for opening the
\href{https://mc-stan.org/}{Stan} progress file (an HTML file). Defaults to option \code{browser} (see
\code{\link[base:options]{base::options()}}). Note that RStudio's default for option \code{browser} cannot be used for the Stan
progress file. Therefore, \pkg{shinybrms} automatically redefines RStudio's default for option
\code{browser} to fix this issue.
\item \code{shinybrms.shinystan_browser}: The value for the global option
\code{browser} to be used for opening the \pkg{shinystan} app. Defaults to option \code{browser} (see
\code{\link[base:options]{base::options()}}). Note that RStudio's default for option \code{browser} cannot be used for opening
the \pkg{shinystan} app from within \pkg{shinybrms}. Therefore, \pkg{shinybrms} automatically
redefines RStudio's default for option \code{browser} to fix this issue.
}

Other global options used by \pkg{shinybrms}:
\itemize{
\item \code{mc.cores} (see \code{\link[base:options]{base::options()}}): If option \code{mc.cores} is set, then the advanced option
"Cores" (in the \pkg{shinybrms} app) defaults to the value of this option. If option \code{mc.cores}
is not set, then the advanced option "Cores" defaults to the value returned by
\code{\link[parallel:detectCores]{parallel::detectCores()}}.
\item \code{shinystan.rstudio} (see \code{\link[shinystan:launch_shinystan]{shinystan::launch_shinystan()}}): If set to
\code{TRUE}, then use RStudio's mechanism for opening the \pkg{shinystan} app (i.e. use
\code{getOption("shiny.launch.browser", interactive())} for argument \code{launch.browser} of
\code{\link[shiny:runApp]{shiny::runApp()}}). Defaults to \code{FALSE} (i.e. use \code{TRUE} for argument \code{launch.browser} of
\code{\link[shiny:runApp]{shiny::runApp()}}).
}
}
\examples{
if(interactive()){
  launch_shinybrms(launch.browser = TRUE)
}
}
